Bug fixes: 0001446: Upon Entering Combat Cursor Changes to Incorrect One Fixed!
0001286: Issues with enemy mob status during/after a fight 
I think I finally nailed this one!

0001411: /Deathblow can kill anybody/thing without any restrictions and causes server crash when destroying vehicles 
You can only deathblow an incapacitated player you are duelling.

+++ src/ZoneServer/WorldManager.cpp
@@ -2720,7 +2720,7 @@ void WorldManager::destroyObject(Object* object)
 
 		default:
 		{
-			gLogger->logMsgF("Unhandled ObjectType in WorldManager::destroyObject: %I64u",MSG_HIGH,object->getType());
+			gLogger->logMsgF("Unhandled ObjectType in WorldManager::destroyObject: %u",MSG_HIGH,(uint32)(object->getType()));
 
 			// Please, when adding new stufff, at least take the time to add a stub for that type.
 			// Better fail always, than have random crashes.
